feat: show "Short text, list" fields on app config page
