+++ apps/commercetools/src/AppConfig/FieldTypeInstructions.tsx
@@ -12,17 +12,19 @@ export const FieldTypeInstructions = ({ contentTypesFound, space, environment }:
     <Heading>Assign to fields</Heading>
     {contentTypesFound ? (
       <Paragraph>
-        This app can only be used with <strong>Short text</strong> fields. Select which fields you’d
-        like to enable for this app.
+        This app can only be used with <strong>Short text</strong> or{' '}
+        <strong>Short text, list</strong> fields. Select which fields you’d like to enable for
+        this app.
       </Paragraph>
     ) : (
       <>
         <Paragraph>
-          This app can only be used with <strong>Short text</strong> fields.
+          This app can only be used with <strong>Short text</strong> or{' '}
+          <strong>Short text, list</strong> fields.
         </Paragraph>
         <Paragraph>
-          There are <strong>no content types with Short text</strong> fields in this environment.
-          You can add one in your{' '}
+          There are <strong>no content types with Short text or Short text, list</strong>{' '}
+          fields in this environment. You can add one in your{' '}
           <TextLink
             linkType="primary"
             target="_blank"
